+++ src/materials/diffusers.rs
@@ -11,11 +11,11 @@ pub fn random_in_hemisphere(normal: &Vec3) -> Vec3 {
 }
 
 pub fn random_in_unit_sphere() -> Vec3 {
-    random_on_unit_sphere() * rand::random::<f64>().sqrt()
+    random_on_unit_sphere() * rand::random::<f32>().sqrt()
 }
 
 pub fn random_on_unit_sphere() -> Vec3 {
-    let (theta, phi): (f64, f64) = rand::random();
+    let (theta, phi): (f32, f32) = rand::random();
 
     Vec3::new(
         theta.sin() * phi.cos(),
@@ -25,11 +25,11 @@ pub fn random_on_unit_sphere() -> Vec3 {
 }
 
 pub fn random_lambertian(normal: &Vec3) -> Vec3 {
-    let (r1, r2): (f64, f64) = rand::random();
+    let (r1, r2): (f32, f32) = rand::random();
     let sin_theta = r1.sqrt();
     let cos_theta = (1.0 - sin_theta * sin_theta).sqrt();
 
-    let phi = 2.0 * std::f64::consts::PI * r2;
+    let phi = 2.0 * std::f32::consts::PI * r2;
 
     let x = phi.cos() * sin_theta;
     let y = phi.sin() * sin_theta;
