+++ src/materials/metal.rs
@@ -5,11 +5,11 @@ use super::{diffusers, Material, ScatterRecord};
 #[derive(Debug, Clone)]
 pub struct Metal {
     albedo: Color,
-    fuzziness: f64,
+    fuzziness: f32,
 }
 
 impl Metal {
-    pub fn new(albedo: Color, fuzziness: f64) -> Self {
+    pub fn new(albedo: Color, fuzziness: f32) -> Self {
         Self { albedo, fuzziness }
     }
 }
