+++ src/ray.rs
@@ -4,7 +4,7 @@ use crate::vec3::{Point3, Vec3};
 pub struct Ray {
     origin: Point3,
     direction: Vec3,
-    time: f64,
+    time: f32,
 }
 
 impl Ray {
@@ -12,7 +12,7 @@ impl Ray {
         Self::new_time_based(origin, direction, 0.)
     }
 
-    pub fn new_time_based(origin: Point3, direction: Vec3, time: f64) -> Self {
+    pub fn new_time_based(origin: Point3, direction: Vec3, time: f32) -> Self {
         Self {
             origin,
             direction,
@@ -28,11 +28,11 @@ impl Ray {
         &self.direction
     }
 
-    pub fn time(&self) -> f64 {
+    pub fn time(&self) -> f32 {
         self.time
     }
 
-    pub fn at(&self, t: f64) -> Point3 {
+    pub fn at(&self, t: f32) -> Point3 {
         self.origin + t * self.direction
     }
 }
