+++ src/textures/image_texture.rs
@@ -15,19 +15,19 @@ impl ImageTexture {
 }
 
 impl Texture for ImageTexture {
-    fn value(&self, u: f64, v: f64, _point: &crate::vec3::Vec3) -> crate::color::Color {
+    fn value(&self, u: f32, v: f32, _point: &crate::vec3::Vec3) -> crate::color::Color {
         let u = u.clamp(0., 1.);
         let v = v.clamp(0., 1.);
 
-        let i = ((u * self.image.width() as f64) as u32).min(self.image.width() - 1);
+        let i = ((u * self.image.width() as f32) as u32).min(self.image.width() - 1);
         let j =
-            (((1. - v) * self.image.height() as f64 - 0.001) as u32).min(self.image.height() - 1);
+            (((1. - v) * self.image.height() as f32 - 0.001) as u32).min(self.image.height() - 1);
 
         let pixel = self.image.get_pixel(i, j);
         Color::new(
-            pixel[0] as f64 / 255.,
-            pixel[1] as f64 / 255.,
-            pixel[2] as f64 / 255.,
+            pixel[0] as f32 / 255.,
+            pixel[1] as f32 / 255.,
+            pixel[2] as f32 / 255.,
         )
     }
 }
