+++ src/textures/noise.rs
@@ -5,11 +5,11 @@ use super::Texture;
 #[derive(Clone)]
 pub struct Noise {
     noise: Perlin,
-    scale: f64,
+    scale: f32,
 }
 
 impl Noise {
-    pub fn new(scale: f64) -> Self {
+    pub fn new(scale: f32) -> Self {
         Self {
             noise: Perlin::new(),
             scale,
@@ -18,7 +18,7 @@ impl Noise {
 }
 
 impl Texture for Noise {
-    fn value(&self, _u: f64, _v: f64, point: &crate::vec3::Vec3) -> crate::color::Color {
+    fn value(&self, _u: f32, _v: f32, point: &crate::vec3::Vec3) -> crate::color::Color {
         Color::new(1., 1., 1.)
             * 0.5
             * (1. + (self.scale * point.z() + 10. * self.noise.turb(point, 7)).sin())
