+++ src/textures/solid_color.rs
@@ -14,7 +14,7 @@ impl SolidColor {
 }
 
 impl Texture for SolidColor {
-    fn value(&self, _u: f64, _v: f64, _point: &Vec3) -> Color {
+    fn value(&self, _u: f32, _v: f32, _point: &Vec3) -> Color {
         self.color
     }
 }
